requirements.txt
streamlit>=1.28.0
ultralytics>=8.0.0
opencv-python-headless>=4.8.1.78,<5.0
pillow>=9.0.0
numpy>=1.21.0